Restructure MMU a bit
