+++ memflow/src/architecture/mmu/spec.rs
@@ -1,6 +1,8 @@
-pub(crate) mod translate_data;
+use super::ArchMMUDef;
+use crate::architecture::Endianess;
 
-use super::Endianess;
+use super::translate_data::{TranslateData, TranslateDataVec, TranslateVec, TranslationChunk};
+use super::MMUTranslationBase;
 use crate::error::{Error, Result};
 use crate::iter::SplitAtIndex;
 use crate::mem::{PhysicalMemory, PhysicalReadData};
@@ -8,7 +10,6 @@ use crate::types::{Address, PageType, PhysicalAddress};
 pub(crate) use fixed_slice_vec::FixedSliceVec as MVec;
 use std::alloc::Layout;
 use std::convert::TryInto;
-use translate_data::{TranslateData, TranslateDataVec, TranslateVec, TranslationChunk};
 
 use bumpalo::Bump;
 
@@ -22,241 +23,6 @@ macro_rules! vtop_trace {
 #[cfg(not(feature = "trace_mmu"))]
 macro_rules! vtop_trace {
     ( $( $x:expr ),* ) => {};
-}
-
-/// The `ArchMMUDef` structure defines how a real memory management unit should behave when
-/// translating virtual memory addresses to physical ones.
-///
-/// The core logic of virtual to physical memory translation is practically the same, but different
-/// MMUs may have different address space sizes, and thus split the addresses in different ways.
-///
-/// For instance, most x86_64 architectures have 4 levels of page mapping, providing 52-bit address
-/// space. Virtual address gets split into 4 9-bit regions, and a 12-bit one, the first 4 are used
-/// to index the page tables, and the last, 12-bit split is used as an offset to get the final
-/// memory address. Meanwhile, x86 with PAE has 3 levels of page mapping, providing 36-bit address
-/// space. Virtual address gets split into a 2-bit, 2 9-bit and a 12-bit regions - the last one is
-/// also used as an offset from the physical frame. The difference is of level count, and virtual
-/// address splits, but the core page table walk stays the same.
-///
-/// Our virtual to physical memory ranslation code is the same for both architectures, in fact, it
-/// is also the same for the x86 (non-PAE) architecture that has different PTE and pointer sizes.
-/// All that differentiates the translation process is the data inside this structure.
-#[derive(Debug)]
-pub struct ArchMMUDef {
-    /// defines the way virtual addresses gets split (the last element
-    /// being the final physical page offset, and thus treated a bit differently)
-    pub virtual_address_splits: &'static [u8],
-    /// defines at which page mapping steps we can return a large page.
-    /// Steps are indexed from 0, and the list has to be sorted, otherwise the code may fail.
-    pub valid_final_page_steps: &'static [usize],
-    /// define the address space upper bound (32 for x86, 52 for x86_64)
-    pub address_space_bits: u8,
-    /// Defines the byte order of the architecture
-    pub endianess: Endianess,
-    /// native pointer size in bytes for the architecture.
-    pub addr_size: u8,
-    /// size of an individual page table entry in bytes.
-    pub pte_size: usize,
-    /// index of a bit in PTE defining whether the page is present or not.
-    pub present_bit: u8,
-    /// index of a bit in PTE defining if the page is writeable.
-    pub writeable_bit: u8,
-    /// index of a bit in PTE defining if the page is non-executable.
-    pub nx_bit: u8,
-    /// index of a bit in PTE defining if the PTE points to a large page.
-    pub large_page_bit: u8,
-}
-
-pub trait MMUTranslationBase: Clone + Copy + core::fmt::Debug {
-    // Retrieves page table address by virtual address
-    fn get_pt_by_virt_addr(&self, address: Address) -> Address;
-
-    // Retreives page table address, and its index by index within
-    // For instance, on ARM index 257 would return kernel page table
-    // address, and index 1. On X86, however, this is a no-op that returns
-    // underlying page table Address and `idx`.
-    fn get_pt_by_index(&self, idx: usize) -> (Address, usize);
-
-    fn pt_count(&self) -> usize;
-
-    fn virt_addr_filter<B: SplitAtIndex, O: Extend<(Error, Address, B)>>(
-        &self,
-        spec: &ArchMMUSpec,
-        addr: (Address, B),
-        work_group: (&mut TranslationChunk<Self>, &mut TranslateDataVec<B>),
-        out_fail: &mut O,
-    );
-
-    fn fill_init_chunk<VI, FO, B>(
-        &self,
-        spec: &ArchMMUSpec,
-        out_fail: &mut FO,
-        addrs: &mut VI,
-        (waiting_addrs, tmp_addrs): (&mut TranslateDataVec<B>, &mut TranslateDataVec<B>),
-        work_vecs: &mut (TranslateVec, TranslateDataVec<B>),
-        next_work_vecs: &mut (TranslateVec, TranslateDataVec<B>),
-    ) where
-        VI: Iterator<Item = (Address, B)>,
-        FO: Extend<(Error, Address, B)>,
-        B: SplitAtIndex,
-    {
-        let mut init_chunk = TranslationChunk::new(*self);
-
-        let working_addr_count = work_vecs.1.capacity();
-
-        for (_, data) in (0..working_addr_count).zip(addrs) {
-            self.virt_addr_filter(spec, data, (&mut init_chunk, waiting_addrs), out_fail);
-            if init_chunk.next_max_addr_count(spec) >= working_addr_count {
-                break;
-            }
-        }
-
-        if init_chunk.addr_count > 0 {
-            vtop_trace!("init_chunk = {:#?}", &init_chunk);
-            init_chunk.split_chunk(spec, (waiting_addrs, tmp_addrs), work_vecs, next_work_vecs);
-        }
-    }
-}
-
-impl MMUTranslationBase for Address {
-    fn get_pt_by_virt_addr(&self, _: Address) -> Address {
-        *self
-    }
-
-    fn get_pt_by_index(&self, idx: usize) -> (Address, usize) {
-        (*self, idx)
-    }
-
-    fn pt_count(&self) -> usize {
-        1
-    }
-
-    fn virt_addr_filter<B, O>(
-        &self,
-        spec: &ArchMMUSpec,
-        addr: (Address, B),
-        work_group: (&mut TranslationChunk<Self>, &mut TranslateDataVec<B>),
-        out_fail: &mut O,
-    ) where
-        B: SplitAtIndex,
-        O: Extend<(Error, Address, B)>,
-    {
-        spec.virt_addr_filter(addr, work_group, out_fail);
-    }
-
-    //TODO: Optimized fill_init_vec impl for non-split page tables
-}
-
-impl ArchMMUDef {
-    pub const fn into_spec(self) -> ArchMMUSpec {
-        ArchMMUSpec::from_def(self)
-    }
-
-    /// Mask a page table entry address to retrieve the next page table entry
-    ///
-    /// This function uses virtual_address_splits to mask the first bits out in `pte_addr`, but
-    /// keep everything else until the `address_space_bits` upper bound.
-    ///
-    /// # Arguments
-    ///
-    /// * `pte_addr` - page table entry address to mask
-    /// * `step` - the current step in the page walk
-    ///
-    /// # Remarks
-    ///
-    /// The final step is handled differently, because the final split provides a byte offset to
-    /// the page, instead of an offset that has to be multiplied by `pte_size`. We do that by
-    /// subtracting `pte_size` logarithm from the split size.
-    pub fn pte_addr_mask(&self, pte_addr: Address, step: usize) -> u64 {
-        let max = self.address_space_bits - 1;
-        let min = self.virtual_address_splits[step]
-            + if step == self.virtual_address_splits.len() - 1 {
-                0
-            } else {
-                self.pte_size.to_le().trailing_zeros() as u8
-            };
-        let mask = Address::bit_mask(min..max);
-        vtop_trace!("pte_addr_mask={:b}", u64::from_le(mask.as_u64()));
-        pte_addr.as_u64() & u64::from_le(mask.as_u64())
-    }
-
-    const fn virt_addr_bit_range(&self, step: usize) -> (u8, u8) {
-        let max_index_bits = {
-            let subsl = &self.virtual_address_splits;
-            let mut accum = 0;
-            let mut i = step;
-            while i < subsl.len() {
-                accum += subsl[i];
-                i += 1;
-            }
-            accum
-        };
-        let min_index_bits = max_index_bits - self.virtual_address_splits[step];
-        (min_index_bits, max_index_bits)
-    }
-
-    /// Return the number of splits of virtual addresses
-    ///
-    /// The returned value will be one more than the number of page table levels
-    pub fn split_count(&self) -> usize {
-        self.virtual_address_splits.len()
-    }
-
-    pub const fn spare_allocs(&self) -> usize {
-        let mut i = 1;
-        let mut fold = 0;
-        while i < self.virtual_address_splits.len() {
-            fold += 1 << self.virtual_address_splits[i - 1];
-            i += 1;
-        }
-        fold
-    }
-
-    /// Calculate the size of the page table entry leaf in bytes
-    ///
-    /// This will return the number of page table entries at a specific step multiplied by the
-    /// `pte_size`. Usually this will be an entire page, but in certain cases, like the highest
-    /// mapping level of x86 with PAE, it will be less.
-    ///
-    /// # Arguments
-    ///
-    /// * `step` - the current step in the page walk
-    pub const fn pt_leaf_size(&self, step: usize) -> usize {
-        let (min, max) = self.virt_addr_bit_range(step);
-        (1 << (max - min)) * self.pte_size
-    }
-
-    /// Get the page size of a specific step without checking if such page could exist
-    ///
-    /// # Arguments
-    ///
-    /// * `step` - the current step in the page walk
-    pub const fn page_size_step_unchecked(&self, step: usize) -> usize {
-        let max_index_bits = {
-            let subsl = &self.virtual_address_splits;
-            let mut i = step;
-            let mut accum = 0;
-            while i < subsl.len() {
-                accum += subsl[i];
-                i += 1;
-            }
-            accum
-        };
-        (1u64 << max_index_bits) as usize
-    }
-
-    /// Get the page size of a specific page walk step
-    ///
-    /// This function is preferable to use externally, because in debug builds it will check if such
-    /// page could exist, and if can not, it will panic
-    ///
-    /// # Arguments
-    ///
-    /// * `step` - the current step in the page walk
-    pub fn page_size_step(&self, step: usize) -> usize {
-        debug_assert!(self.valid_final_page_steps.binary_search(&step).is_ok());
-        self.page_size_step_unchecked(step)
-    }
 }
 
 const MAX_LEVELS: usize = 8;
@@ -403,6 +169,7 @@ impl ArchMMUSpec {
         }
     }
 
+    #[allow(unused)]
     pub fn split_count(&self) -> usize {
         self.def.virtual_address_splits.len()
     }
